move date validation to server
