+++ stats-service/stats-service-gateway/src/main/java/ru/practicum/explorewithme/gateway/error/controller/ErrorHandler.java
@@ -7,19 +7,12 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import ru.practicum.explorewithme.gateway.error.model.ErrorResponse;
-import ru.practicum.explorewithme.gateway.exceptions.ValidationException;
 
 import java.util.Objects;
 
 @Slf4j
 @RestControllerAdvice(basePackages = "ru.practicum.explorewithme")
 public class ErrorHandler {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleValidated(final ValidationException e) {
-        return new ErrorResponse(e.getMessage());
-    }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
