--- stats-service/stats-service-gateway/src/main/java/ru/practicum/explorewithme/gateway/exceptions/ValidationException.java
@@ -1,7 +1,0 @@
-package ru.practicum.explorewithme.gateway.exceptions;
-
-public class ValidationException extends RuntimeException {
-    public ValidationException(String message) {
-        super(message);
-    }
-}
