+++ stats-service/stats-service-gateway/src/main/java/ru/practicum/explorewithme/gateway/stats/StatsClient.java
@@ -9,22 +9,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.util.DefaultUriBuilderFactory;
 import ru.practicum.explorewithme.dto.stats.EndpointHitRequestDto;
 import ru.practicum.explorewithme.gateway.client.BaseClient;
-import ru.practicum.explorewithme.gateway.exceptions.ValidationException;
 
-import java.net.URLDecoder;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 @Service
 public class StatsClient extends BaseClient {
-
-    public static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
     @Value("${stats-server.hit}")
     private String hitUrl;
@@ -48,14 +41,6 @@ public class StatsClient extends BaseClient {
     }
 
     public ResponseEntity<Object> getStats(String start, String end, List<String> uris, Boolean unique) {
-        try {
-            start = LocalDateTime.parse(URLDecoder.decode(start, StandardCharsets.UTF_8), FORMATTER)
-                    .format(DateTimeFormatter.ISO_DATE_TIME);
-            end = LocalDateTime.parse(URLDecoder.decode(end, StandardCharsets.UTF_8), FORMATTER)
-                    .format(DateTimeFormatter.ISO_DATE_TIME);
-        } catch (DateTimeParseException | IllegalArgumentException e) {
-            throw new ValidationException("Период времени необходимо передавать в формате 'yyyy-MM-dd HH:mm:ss'");
-        }
         Map<String, Object> parameters = new HashMap<>(Map.of(
                 "start", URLEncoder.encode(start, StandardCharsets.UTF_8),
                 "end", URLEncoder.encode(end, StandardCharsets.UTF_8),
