+++ stats-service/stats-service-server/src/main/java/ru/practicum/explorewithme/server/stats/controller/StatsController.java
@@ -32,7 +32,7 @@ public class StatsController {
     public List<ViewStatsDto> getStatsByPeriodAndUris(@RequestParam String start,
                                                       @RequestParam String end,
                                                       @RequestParam(required = false) List<String> uris,
-                                                      @RequestParam Boolean unique) {
+                                                      @RequestParam(defaultValue = "false") Boolean unique) {
         return statsService.getStatsByPeriodAndUris(start, end, uris, unique);
     }
 }
