+++ stats-service/stats-service-server/src/main/java/ru/practicum/explorewithme/server/stats/service/StatsServiceImpl.java
@@ -15,6 +15,8 @@ import ru.practicum.explorewithme.server.stats.model.ViewStats;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 import java.util.List;
 
 @Slf4j
@@ -24,6 +26,7 @@ import java.util.List;
 public class StatsServiceImpl implements StatsService {
 
     private final StatsRepository statsRepository;
+    public static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
     @Override
     @Transactional
@@ -36,8 +39,16 @@ public class StatsServiceImpl implements StatsService {
     @Override
     public List<ViewStatsDto> getStatsByPeriodAndUris(String start, String end, List<String> uris, Boolean unique) {
         log.info("Запрос на получение статистик для периода {}-{}, по uri {} с уникальностью = {}", start, end, uris, unique);
-        LocalDateTime startDate = LocalDateTime.parse(URLDecoder.decode(start, StandardCharsets.UTF_8));
-        LocalDateTime endDate = LocalDateTime.parse(URLDecoder.decode(end, StandardCharsets.UTF_8));
+        try {
+            start = LocalDateTime.parse(URLDecoder.decode(start, StandardCharsets.UTF_8), FORMATTER)
+                    .format(DateTimeFormatter.ISO_DATE_TIME);
+            end = LocalDateTime.parse(URLDecoder.decode(end, StandardCharsets.UTF_8), FORMATTER)
+                    .format(DateTimeFormatter.ISO_DATE_TIME);
+        } catch (DateTimeParseException | IllegalArgumentException e) {
+            throw new ValidationException("Период времени необходимо передавать в формате 'yyyy-MM-dd HH:mm:ss'");
+        }
+        LocalDateTime startDate = LocalDateTime.parse(start);
+        LocalDateTime endDate = LocalDateTime.parse(end);
         if (endDate.isBefore(startDate) || endDate.equals(startDate)) {
             throw new ValidationException("Даты не могут быть равны или дата окончания не может быть раньше даты начала");
         }
